snapcheck-recognize-Model-Backend/requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy
pandas
matplotlib
scikit-learn
opencv-python
aiohttp
aiohttp-session
aiortc
python-dotenv
rich
sqlalchemy
pytz
pytest
mediapipe
kaggle
schedule
pymysql